pyproject.toml
[project]
name = "smart-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.1.1",
    "dashscope>=1.24.5",
    "fastapi>=0.116.2",
    "uvicorn[standard]>=0.35.0",
    "python-socketio>=5.13.0",
    "jinja2>=3.1.6",
    "openai>=1.109.0",
    "dify-client>=0.1.10",
    "sqlalchemy[asyncio]>=2.0.48",
    "asyncpg>=0.31.0",
    "redis>=7.3.0",
    "fastmcp>=3.1.0",
]

[[tool.uv.index]]
#url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
url = "https://maven.bsoftpay.com/repository/tsinghua/simple/"
default = true
